+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [v0.11.0.0](https://github.com/LambdaHack/LambdaHack/compare/v0.10.3.0...v0.11.0.0)
 
+- Split the long Content.ModeKind texts into lines, [issue 283](https://github.com/LambdaHack/LambdaHack/issues/283)
 - Partially work around regression in libsdl2 2.0.16 (https://github.com/LambdaHack/LambdaHack/issues/281); to also avoid deformed boxes around tiles on the game map, please switch to a different SDL2 version
 - Deduplicate UI code for exiting game, with extra style points
 - Create monadic test harness and use it for UI and other unit tests